pyproject.toml
[project]
name = "linux-gpio-http-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "David Lechner", email = "dlechner@baylibre.com" }
]
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "gpiod>=2.4.0",
    "pyudev>=0.24.4",
    "ruamel-yaml>=0.19.1",
]

[project.scripts]
linux-gpio-http-server = "linux_gpio_http_server:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]

[tool.ruff]
target-version = "py312"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
